Fix jira NCF-243
Problem was that the NC_create
code was not checking for the NC_CLASSIC_MODEL
mode flag in deciding what dispatch table to use.
This meant that it was then defaulting to use
the default format, and if that was changed
to e.g. NC_FORMAT_NETCDF4, then it would try
to create a netcdf-4 format file, even is
NC_CLASSIC_MODEL mode flag was set.


git-svn-id: https://sub.unidata.ucar.edu/netcdf/trunk@3069 2e3060e4-1ed8-cf1f-9cc2-9586e844a556